Small structure changes and some fixes
Added a timeout property to each message type, to make it possible
to send a message only once in a while. The timeout is given in
milliseconds.

All message types has a default timeout of 3000 ms, and some messages
are given a higher timeout to make them not trigger as often.

Added a new message type that doesn't trigger on messages, but
instead are checked to see if it should run every n seconds (given
by the INTERVAL variable).

Tweaking the anger detector a bit.